Streamlit_Dashboards/requirements.txt
streamlit
reportlab
pandas
mysql-connector-python
groq
altair
smolagents
numpy
scikit-learn
streamlit-folium
folium
urllib3
asyncio
pydantic-ai
datetime
scikit-learn
reportlab
svglib

